refactor: change name to scatter from embed
