+++ frontend/src/instance-views/ScatterView.svelte
